+++ extensions/battery-menubar/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Battery Menu Bar Changelog
 
+## [Added New Features] - 2023-09-30
+
+- Add battery cycle count
+- Add battery health
+
 ## [Initial Version] - 2023-08-08